fix executeCommandLine without a Duration
Without a `Duration` argument, `executeCommandLine` always returns `null`. I removed the example that appears to show calling it without a `Duration` and getting the result of the echo.  That can't happen as it's currently implemented.

Also, we need to use `time.Duration` right? I'm pretty sure that's the case. If I'm wrong and that's not the case, we should show importing java.time.Duration in the example.