Markdown Checks: Upgrade actions & Ignore commit/PR links (#436)
This updates the Markdown Broken Link Check to use
tcort/github-action-markdown-link-check instead of the deprecated
garauv-nelson/github-action-markdown-link-check and pins its version to
a SHA to migitate possible supply chain attacks.
It also ignores commit/PR links for this repository.

This updates the markdownlint action and pins its version to a SHA.

---------

Signed-off-by: Florian Hotze <dev@florianhotze.com>