[rules] Expose _getTriggeredData & Adjust for backward compat with UI scripts (#466)
Closes #443.

- Exposes the `_getTriggeredData` utility from the `rules` namespace,
which converts the Java event data to pure JS.
- Adds optional backwards compatibilty with the pure Java event object
in UI-based environments by adding its event properties to the event
object if a flag is set. Accessing these properties logs warnings.
- One **minor breaking change** for UI-based environments: The
`.payload` property of `_getTriggeredData` returns a object, the pure
Java version returns a JSON string that has to be parsed by the user.

Signed-off-by: Florian Hotze <dev@florianhotze.com>